backend/requirements.txt
MetaTrader5==5.0.45
pandas==2.2.1
pyyaml==6.0.1
loguru==0.7.2
python-dotenv==1.0.1
